+++ include/stumpless/target.h
@@ -31,7 +31,8 @@ extern "C" {
   enum stumpless_target_type {
     STUMPLESS_BUFFER_TARGET,
     STUMPLESS_FILE_TARGET,
-    STUMPLESS_SOCKET_TARGET
+    STUMPLESS_SOCKET_TARGET,
+    STUMPLESS_WINDOWS_EVENT_LOG_TARGET
   };
 
   struct stumpless_target {