+++ src/error.c
@@ -77,6 +77,7 @@ void
 raise_invalid_id( void ) {
   raise_error( STUMPLESS_INVALID_ID );
 }
+
 void
 raise_memory_allocation_failure( void ) {
   raise_error( STUMPLESS_MEMORY_ALLOCATION_FAILURE );
@@ -90,4 +91,14 @@ raise_socket_bind_failure( void ) {
 void
 raise_target_unsupported( void ) {
   raise_error( STUMPLESS_TARGET_UNSUPPORTED );
+}
+
+void
+raise_wel_close_failure( void ) {
+  raise_error( STUMPLESS_WINDOWS_EVENT_LOG_CLOSE_FAILURE );
+}
+
+void
+raise_wel_open_failure( void ) {
+  raise_error( STUMPLESS_WINDOWS_EVENT_LOG_OPEN_FAILURE );
 }