+++ test/function/entry.cpp
@@ -16,8 +16,10 @@
  * limitations under the License.
  */
 
-#include <stddef.h>
 #include <gtest/gtest.h>
+#include <stddef.h>
+#include <stdlib.h>
+#include <string.h>
 #include <stumpless.h>
 
 namespace {
@@ -68,14 +70,14 @@ namespace {
     struct stumpless_entry *entry;
     struct stumpless_element *element;
     struct stumpless_error *error;
-    void *(*result)(size_t);
+    void * (*set_realloc_result)(void *, size_t);
 
     element = stumpless_new_element( "test-memory-failure" );
     ASSERT_TRUE( element != NULL );
     EXPECT_EQ( NULL, stumpless_get_error(  ) );
    
-    result = stumpless_set_malloc( [](size_t size)->void *{ return NULL; } );
-    ASSERT_TRUE( result != NULL );
+    set_realloc_result = stumpless_set_realloc( [](void *, size_t)->void *{ return NULL; } );
+    ASSERT_TRUE( set_realloc_result != NULL );
  
     entry = stumpless_add_element( basic_entry, element );
     EXPECT_EQ( NULL, entry );
@@ -87,7 +89,8 @@ namespace {
       EXPECT_EQ( error->id, STUMPLESS_MEMORY_ALLOCATION_FAILURE );
     }
 
-    stumpless_set_malloc( malloc );
+    set_realloc_result = stumpless_set_realloc( realloc );
+    EXPECT_TRUE( set_realloc_result == realloc );
   }
   
   TEST_F( EntryTest, AddNullElement ) {
@@ -548,9 +551,4 @@ namespace {
     EXPECT_EQ( error->id, STUMPLESS_ARGUMENT_EMPTY );
   }
 
-}
-
-int main( int argc, char **argv ) {
-  ::testing::InitGoogleTest( &argc, argv );
-  return RUN_ALL_TESTS(  );
 }