+++ test/function/startup/target/socket_add_malloc_failure.cpp
@@ -94,9 +94,3 @@ namespace {
   }
 
 }
-
-int
-main( int argc, char **argv ) {
-  ::testing::InitGoogleTest( &argc, argv );
-  return RUN_ALL_TESTS(  );
-}