+++ test/function/target.cpp
@@ -117,8 +117,3 @@ namespace {
     stumpless_close_buffer_target( target );
   }
 }
-
-int main(int argc, char **argv){
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}