+++ test/function/target/buffer.cpp
@@ -240,9 +240,3 @@ namespace {
     }
   }
 }
-
-int
-main( int argc, char **argv ) {
-  ::testing::InitGoogleTest( &argc, argv );
-  return RUN_ALL_TESTS(  );
-}