+++ tools/cmake/gtest.cmake
@@ -1,8 +1,29 @@
+set(my_gtest_binary_dir "${CMAKE_CURRENT_BINARY_DIR}/gtest/src/gtest-build")
+
+if(WIN32)
+  if(CMAKE_BUILD_TYPE STREQUAL "Debug")
+    set(my_gtest_imported_location "${my_gtest_binary_dir}/lib/${CMAKE_CFG_INTDIR}/gtestd.lib")
+    set(my_gtest_main_imported_location "${my_gtest_binary_dir}/lib/${CMAKE_CFG_INTDIR}/gtest_maind.lib")
+    set(my_gmock_imported_location "${my_gtest_binary_dir}/lib/${CMAKE_CFG_INTDIR}/gmockd.lib")
+  else()
+    set(my_gtest_imported_location "${my_gtest_binary_dir}/lib/${CMAKE_CFG_INTDIR}/gtest.lib")
+    set(my_gtest_main_imported_location "${my_gtest_binary_dir}/lib/${CMAKE_CFG_INTDIR}/gtest_main.lib")
+    set(my_gmock_imported_location "${my_gtest_binary_dir}/lib/${CMAKE_CFG_INTDIR}/gmock.lib")
+  endif(CMAKE_BUILD_TYPE STREQUAL "Debug")
+else()
+  set(my_gtest_imported_location "${my_gtest_binary_dir}/lib/libgtest.a")
+  set(my_gtest_main_imported_location "${my_gtest_binary_dir}/lib/libgtest_main.a")
+  set(my_gmock_imported_location "${my_gtest_binary_dir}/lib/libgmock.a")
+endif(WIN32)
+
 ExternalProject_Add(gtest
-    URL https://github.com/google/googletest/archive/master.zip
-    PREFIX ${CMAKE_CURRENT_BINARY_DIR}/gtest
-    CMAKE_ARGS -Dgtest_force_shared_crt=ON
-    INSTALL_COMMAND ""
+  URL https://github.com/google/googletest/archive/master.zip
+  PREFIX ${CMAKE_CURRENT_BINARY_DIR}/gtest
+  CMAKE_ARGS -Dgtest_force_shared_crt=ON
+  UPDATE_COMMAND ""
+  INSTALL_COMMAND ""
+  BINARY_DIR "${my_gtest_binary_dir}"
+  BUILD_BYPRODUCTS "${my_gtest_imported_location}" "${my_gtest_main_imported_location}" "${my_gmock_imported_location}"
 )
 
 set_target_properties(gtest
@@ -10,44 +31,34 @@ set_target_properties(gtest
   EXCLUDE_FROM_ALL TRUE
 )
 
-ExternalProject_Get_Property(gtest source_dir binary_dir)
-
 add_library(libgtest IMPORTED STATIC GLOBAL)
 add_dependencies(libgtest gtest)
+set(gtest_imported_location "${my_gtest_imported_location}")
 
-if(WIN32)
-  if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-    set(gtest_imported_location "${binary_dir}/googlemock/gtest/${CMAKE_CFG_INTDIR}/gtestd.lib")
-  else()
-    set(gtest_imported_location "${binary_dir}/googlemock/gtest/${CMAKE_CFG_INTDIR}/gtest.lib")
-  endif(CMAKE_BUILD_TYPE STREQUAL "Debug")
-else()
-  set(gtest_imported_location "${binary_dir}/googlemock/gtest/libgtest.a")
-endif(WIN32)
+add_library(libgtestmain IMPORTED STATIC GLOBAL)
+add_dependencies(libgtestmain gtest)
+set(gtest_main_imported_location "${my_gtest_main_imported_location}")
+
+add_library(libgmock IMPORTED STATIC GLOBAL)
+add_dependencies(libgmock gtest)
+set(gmock_imported_location "${my_gmock_imported_location}")
 
 set_target_properties(libgtest PROPERTIES
   IMPORTED_LOCATION "${gtest_imported_location}"
   IMPORTED_LINK_INTERFACE_LIBRARIES "${CMAKE_THREAD_LIBS_INIT}"
 )
 
-add_library(libgmock IMPORTED STATIC GLOBAL)
-add_dependencies(libgmock gtest)
-
-if(WIN32)
-  if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-    set(gmock_imported_location "${binary_dir}/googlemock/${CMAKE_CFG_INTDIR}/gmockd.lib")
-  else()
-    set(gmock_imported_location "${binary_dir}/googlemock/${CMAKE_CFG_INTDIR}/gmock.lib")
-  endif(CMAKE_BUILD_TYPE STREQUAL "Debug")
-else()
- set(gmock_imported_location "${binary_dir}/googlemock/libgmock.a")
-endif(WIN32)
+set_target_properties(libgtestmain PROPERTIES
+  IMPORTED_LOCATION "${gtest_main_imported_location}"
+  IMPORTED_LINK_INTERFACE_LIBRARIES "${CMAKE_THREAD_LIBS_INIT}"
+)
 
 set_target_properties(libgmock PROPERTIES
   IMPORTED_LOCATION "${gmock_imported_location}"
   IMPORTED_LINK_INTERFACE_LIBRARIES "${CMAKE_THREAD_LIBS_INIT}"
 )
 
+ExternalProject_Get_Property(gtest source_dir)
 include_directories("${source_dir}/googletest/include"
                     "${source_dir}/googlemock/include"
 )