+++ ChangeLog.md
@@ -5,16 +5,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
- - Add windows event log target capability.
+ - Network target capability.
+ - Compile-time log level filtering.
+ - Format-string style logging from `stumpless` call.
+ - Thread-safe operation.
+ - Examples for file and socket targets.
 
-## [1.2.0] - 2018-09-15
+## [1.2.0] - 2018-11-3
 ### Added
  - Increased coverage on object cache.
  - Support for fractional timestamps.
  - File target capability.
+ - Windows Event Log target capability.
+ - Doxygen documentation support.
 
 ### Removed
  - Use of unsafe gmtime function.
+
+## [1.1.2] - 2018-10-04
+### Fixed
+ - Specify googletest release 1.8.1 to fix broken builds.
 
 ## [1.1.1] - 2018-08-28
 ### Added